# Conflicts:
#	backend/src/main/java/com/example/modules/user/UserService.java
#	backend/src/main/java/com/example/modules/user/web/UserController.java
#	backend/src/main/java/com/example/shared/ApiHttpResponse.java
#	backend/src/main/resources/application.yml

+++ backend/src/main/java/com/example/modules/user/UserService.java
@@ -1,10 +1,14 @@
 package com.example.modules.user;
 
 import com.example.modules.user.web.UserDTO;
+
+import com.example.modules.user.web.UserReadDTO;
 import com.example.shared.CrudService;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
+import org.modelmapper.ModelMapper;
 import org.springframework.stereotype.Service;
+import org.springframework.http.HttpStatus;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
@@ -15,6 +19,7 @@ public class UserService implements CrudService<UserDTO> {
     
     private final UserRepository userRepository;
     private final UserMapper userMapper;
+    private final ModelMapper modelMapper;
     
     public List<UserDTO> getAll() {
         return userRepository.findAll().stream().map(userMapper::toDto).toList();
@@ -22,7 +27,14 @@ public class UserService implements CrudService<UserDTO> {
     
     @Override
     public UserDTO get(Long id) {
-        return userMapper.toDto(userRepository.findById(id).orElseThrow());
+        User user = userRepository.findById(id).orElseThrow();
+        return modelMapper.map(user, UserDTO.class);
+        //        return userMapper.toDto(userRepository.findById(id).orElseThrow());
+    }
+    
+    public UserReadDTO getSmallDTO(Long id) {
+        User user = userRepository.findById(id).orElseThrow();
+        return modelMapper.map(user, UserReadDTO.class);
     }
     
     @Override
@@ -33,9 +45,10 @@ public class UserService implements CrudService<UserDTO> {
         return userMapper.toDto(user);
     }
     
+    @Transactional
     @Override
     public UserDTO update(UserDTO userDTO) {
-        User user = userRepository.getReferenceById(userDTO.getUser_id());
+        User user = userRepository.getReferenceById(userDTO.getId());
         userMapper.toEntity(userDTO, user); // apply changes
         user = userRepository.save(user);
         return userMapper.toDto(user);
@@ -43,10 +56,8 @@ public class UserService implements CrudService<UserDTO> {
     
     @Override
     public void delete(Long id) {
-        if (userRepository.existsById(id)) {
-            userRepository.deleteById(id);
-        }
-        else throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Couldn't find the provided user.");
+        if (userRepository.existsById(id)) userRepository.deleteById(id);
+        else throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Couldn't find the user.");
     }
     
 //    @PostConstruct
